+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - AIDER_DESK_DATA_DIR now properly respected
 - added support for ARGUMENTS placeholder in custom commands
 - reloading project keeps the current session
+- improved performance of model library by using virtualized table for models
 
 ## [0.33.0]